Backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-cloud-vision==2.4.2
openai==0.27.0